Revert "More deprecation fix"
This reverts commit f5f25f61d8d670fa4c3b6a24442e49adaba7bfb1.

+++ Ghidra/Extensions/sample/src/main/java/ghidra/examples/EntropyFieldFactory.java
@@ -108,7 +108,7 @@ public class EntropyFieldFactory extends FieldFactory {
 			if (count == 0) {
 				continue;
 			}
-			double p_x = (double) count / (double) len;
+			double p_x = (double) count / len;
 			sum -= p_x * Math.log(p_x) * 1.0 / Math.log(2);
 		}
