+++ Ghidra/Framework/SoftwareModeling/src/main/java/ghidra/program/database/references/BigRefListV0.java
@@ -364,7 +364,7 @@ class BigRefListV0 extends RefList {
 	}
 
 	private byte findHighestRefLevel(byte currentRefLevel) throws IOException {
-		byte maxLevel = (byte) -1;
+		byte maxLevel = -1;
 		ReferenceIterator it = getRefs();
 		while (it.hasNext()) {
 			Reference ref = it.next();
